// frontend/src/app/core/models/role.enum.ts
export enum Role {
  VENDEUR = 'VENDEUR',
  TRANSPORTEUR = 'TRANSPORTEUR',
  ACHETEUR = 'ACHETEUR',
  VETERINAIRE = 'VETERINAIRE',
  // Valeur alignée sur l'enum backend (com.marketplace.model.Role.ANADER)
  AGENT_ANADER = 'ANADER',
  ADMINISTRATEUR = 'ADMINISTRATEUR',
  ADMIN = 'ADMIN',
  USER = 'USER',
}